Added README, streams example+ quick refactor
