+++ clients/cloud/java-springboot/src/main/resources/application-ccloud.properties
@@ -24,4 +24,12 @@ spring.kafka.producer.value-serializer=io.confluent.kafka.serializers.KafkaAvroS
 # consumer configuration 
 spring.kafka.consumer.group-id=java-springboot
 spring.kafka.consumer.key-deserializer=org.apache.kafka.common.serialization.StringDeserializer
+spring.kafka.consumer.value-deserializer=io.confluent.kafka.serializers.KafkaAvroDeserializerspring.kafka.consumer.value-deserializer=io.confluent.kafka.serializers.KafkaAvroDeserializer
 spring.kafka.consumer.value-deserializer=io.confluent.kafka.serializers.KafkaAvroDeserializer
+
+# kafka streams properties
+spring.kafka.streams.application-id=count-aggregator
+spring.kafka.streams.properties.default.key.serde=org.apache.kafka.common.serialization.Serdes$StringSerde
+spring.kafka.streams.properties.default.value.serde=io.confluent.kafka.streams.serdes.avro.SpecificAvroSerde
+spring.kafka.streams.properties.commit.interval.ms=0
+spring.kafka.streams.properties.replication.factor=3
