+++ clients/cloud/java-springboot/src/main/resources/application.properties
@@ -11,8 +11,15 @@ spring.kafka.consumer.properties.specific.avro.reader=true
 # producer config 
 spring.kafka.producer.key-serializer=org.apache.kafka.common.serialization.StringSerializer
 spring.kafka.producer.value-serializer=io.confluent.kafka.serializers.KafkaAvroSerializer
-
+￿
 # consumer config
 spring.kafka.consumer.group-id=test
 spring.kafka.consumer.key-deserializer=org.apache.kafka.common.serialization.StringDeserializer
 spring.kafka.consumer.value-deserializer=io.confluent.kafka.serializers.KafkaAvroDeserializer
+        
+# kafka streams properties
+spring.kafka.streams.application-id=count-aggregator
+spring.kafka.streams.properties.default.key.serde=org.apache.kafka.common.serialization.Serdes$StringSerde
+spring.kafka.streams.properties.default.value.serde=io.confluent.kafka.streams.serdes.avro.SpecificAvroSerde
+spring.kafka.streams.properties.commit.interval.ms=0
+spring.kafka.streams.properties.replication.factor=1
